Fixed the error that was causing extra new lines to be added into the accountsFile.txt. Left error handling in just in case.
Fixed an error in the account creation process where the password was being forced to have a special character as opposed to not being able to have any.

Added functionality to the username/account system. Only admins can delete resources and access the account management system.

Added account management system. Prints all account and their levels, allows for promotion and demotion of admin level, and allows for deletion of accounts.